+++ package/src/external/reanimated/index.ts
@@ -3,3 +3,4 @@ export * from "./useDerivedValueOnJS";
 export * from "./renderHelpers";
 export * from "./interpolators";
 export * from "./textures";
+export * from "./buffers";
